+++ command/deploy.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/hashicorp/levant/helper"
+	"github.com/hashicorp/levant/levant"
+	"github.com/hashicorp/levant/levant/structs"
+	"github.com/hashicorp/levant/logging"
+	"github.com/hashicorp/levant/template"
 	nomad "github.com/hashicorp/nomad/api"
-	"github.com/jrasell/levant/helper"
-	"github.com/jrasell/levant/levant"
-	"github.com/jrasell/levant/levant/structs"
-	"github.com/jrasell/levant/logging"
-	"github.com/jrasell/levant/template"
 )
 
 // DeployCommand is the command implementation that allows users to deploy a
@@ -136,7 +136,7 @@ func (c *DeployCommand) Run(args []string) int {
 
 	args = flags.Args()
 
-	if config.Deploy.EnvVault == true && config.Deploy.VaultToken != "" {
+	if config.Deploy.EnvVault && config.Deploy.VaultToken != "" {
 		c.UI.Error(c.Help())
 		c.UI.Error("\nERROR: Can not used -vault and -vault-token flag at the same time")
 		return 1