+++ helper/variable.go
@@ -7,7 +7,7 @@ import (
 // VariableMerge merges the passed file variables with the flag variabes to
 // provide a single set of variables. The flagVars will always prevale over file
 // variables.
-func VariableMerge(fileVars *map[string]interface{}, flagVars *map[string]string) map[string]interface{} {
+func VariableMerge(fileVars, flagVars *map[string]interface{}) map[string]interface{} {
 
 	out := make(map[string]interface{})