+++ template/funcs.go
@@ -11,7 +11,11 @@ import (
 	"strings"
 	"text/template"
 	"time"
+	"unicode"
+	"unicode/utf8"
 
+	"github.com/Masterminds/sprig/v3"
+	spewLib "github.com/davecgh/go-spew/spew"
 	consul "github.com/hashicorp/consul/api"
 	"github.com/rs/zerolog/log"
 )
@@ -19,7 +23,7 @@ import (
 // funcMap builds the template functions and passes the consulClient where this
 // is required.
 func funcMap(consulClient *consul.Client) template.FuncMap {
-	return template.FuncMap{
+	r := template.FuncMap{
 		"consulKey":          consulKeyFunc(consulClient),
 		"consulKeyExists":    consulKeyExistsFunc(consulClient),
 		"consulKeyOrDefault": consulKeyOrDefaultFunc(consulClient),
@@ -31,6 +35,7 @@ func funcMap(consulClient *consul.Client) template.FuncMap {
 		"parseInt":           parseInt,
 		"parseJSON":          parseJSON,
 		"parseUint":          parseUint,
+		"replace":            replace,
 		"timeNow":            timeNowFunc,
 		"timeNowUTC":         timeNowUTCFunc,
 		"timeNowTimezone":    timeNowTimezoneFunc(),
@@ -43,6 +48,44 @@ func funcMap(consulClient *consul.Client) template.FuncMap {
 		"multiply": multiply,
 		"divide":   divide,
 		"modulo":   modulo,
+
+		// Case Helpers
+		"firstRuneToUpper": firstRuneToUpper,
+		"firstRuneToLower": firstRuneToLower,
+		"runeToUpper":      runeToUpper,
+		"runeToLower":      runeToLower,
+
+		//debug.
+		"spewDump":   spewDump,
+		"spewPrintf": spewPrintf,
+	}
+	// Add the Sprig functions to the funcmap
+	for k, v := range sprig.FuncMap() {
+		// if there is a name conflict, favor sprig and rename original version
+		if origFun, ok := r[k]; ok {
+			if name, err := firstRuneToUpper(k); err == nil {
+				name = "levant" + name
+				log.Debug().Msgf("template/funcs: renaming \"%v\" function to \"%v\"", k, name)
+				r[name] = origFun
+			} else {
+				log.Error().Msgf("template/funcs: could not add \"%v\" function. error:%v", k, err)
+			}
+		}
+		r[k] = v
+	}
+	r["sprigVersion"] = sprigVersionFunc
+
+	return r
+}
+
+// SprigVersion contains the semver of the included sprig library
+// it is used in command/version and provided in the sprig_version
+// template function
+const SprigVersion = "3.1.0"
+
+func sprigVersionFunc() func(string) (string, error) {
+	return func(s string) (string, error) {
+		return SprigVersion, nil
 	}
 }
 
@@ -200,6 +243,10 @@ func parseUint(s string) (uint64, error) {
 		return 0, err
 	}
 	return result, nil
+}
+
+func replace(input, from, to string) string {
+	return strings.Replace(input, from, to, -1)
 }
 
 func timeNowFunc() string {
@@ -454,4 +501,48 @@ func modulo(b, a interface{}) (interface{}, error) {
 	default:
 		return nil, fmt.Errorf("modulo: unknown type for %q (%T)", av, a)
 	}
+}
+
+func firstRuneToUpper(s string) (string, error) {
+	return runeToUpper(s, 0)
+}
+
+func runeToUpper(inString string, runeIndex int) (string, error) {
+	return funcOnRune(unicode.ToUpper, inString, runeIndex)
+}
+
+func firstRuneToLower(s string) (string, error) {
+	return runeToLower(s, 0)
+}
+
+func runeToLower(inString string, runeIndex int) (string, error) {
+	return funcOnRune(unicode.ToLower, inString, runeIndex)
+}
+
+func funcOnRune(inFunc func(rune) rune, inString string, runeIndex int) (string, error) {
+	if !utf8.ValidString(inString) {
+		return "", errors.New("funcOnRune: not a valid UTF-8 string")
+	}
+
+	runeCount := utf8.RuneCountInString(inString)
+
+	if runeIndex > runeCount-1 || runeIndex < 0 {
+		return "", fmt.Errorf("funcOnRune: runeIndex out of range (max:%v, provided:%v)", runeCount-1, runeIndex)
+	}
+	runes := []rune(inString)
+	transformedRune := inFunc(runes[runeIndex])
+
+	if runes[runeIndex] == transformedRune {
+		return inString, nil
+	}
+	runes[runeIndex] = transformedRune
+	return string(runes), nil
+}
+
+func spewDump(a interface{}) (string, error) {
+	return spewLib.Sdump(a), nil
+}
+
+func spewPrintf(format string, args ...interface{}) (string, error) {
+	return spewLib.Sprintf(format, args), nil
 }