+++ template/render.go
@@ -7,32 +7,31 @@ import (
 	"io/ioutil"
 	"path"
 
-	"github.com/jrasell/levant/client"
-	"github.com/jrasell/levant/helper"
-	"github.com/rs/zerolog/log"
-	yaml "gopkg.in/yaml.v2"
-
+	"github.com/hashicorp/levant/client"
+	"github.com/hashicorp/levant/helper"
 	nomad "github.com/hashicorp/nomad/api"
 	"github.com/hashicorp/nomad/jobspec"
-	"github.com/hashicorp/terraform/config"
+	"github.com/hashicorp/terraform/configs"
+	"github.com/hashicorp/terraform/configs/hcl2shim"
+	"github.com/rs/zerolog/log"
+	yaml "gopkg.in/yaml.v2"
 )
 
 // RenderJob takes in a template and variables performing a render of the
 // template followed by Nomad jobspec parse.
-func RenderJob(templateFile string, variableFiles []string, addr string, strictMode bool, flagVars *map[string]string) (job *nomad.Job, err error) {
+func RenderJob(templateFile string, variableFiles []string, addr string, strictMode bool, flagVars *map[string]interface{}) (job *nomad.Job, err error) {
 	var tpl *bytes.Buffer
 	tpl, err = RenderTemplate(templateFile, variableFiles, addr, strictMode, flagVars)
 	if err != nil {
 		return
 	}
 
-	job, err = jobspec.Parse(tpl)
-	return
+	return jobspec.Parse(tpl)
 }
 
 // RenderTemplate is the main entry point to render the template based on the
 // passed variables file.
-func RenderTemplate(templateFile string, variableFiles []string, addr string, strictMode bool, flagVars *map[string]string) (tpl *bytes.Buffer, err error) {
+func RenderTemplate(templateFile string, variableFiles []string, addr string, strictMode bool, flagVars *map[string]interface{}) (tpl *bytes.Buffer, err error) {
 
 	t := &tmpl{}
 	t.flagVariables = flagVars
@@ -47,18 +46,16 @@ func RenderTemplate(templateFile string, variableFiles []string, addr string, st
 
 	t.consulClient = c
 
-	if len(variableFiles) == 0 {
+	if len(t.variableFiles) == 0 {
 		log.Debug().Msgf("template/render: no variable file passed, trying defaults")
-		defaultVarFile := helper.GetDefaultVarFile()
-		if defaultVarFile != "" {
-			t.variableFiles = make([]string, 1)
-			t.variableFiles[0] = defaultVarFile
-			log.Debug().Msgf("template/render: found default variable file, using %s", t.variableFiles[0])
+		if defaultVarFile := helper.GetDefaultVarFile(); defaultVarFile != "" {
+			t.variableFiles = []string{defaultVarFile}
+			log.Debug().Msgf("template/render: found default variable file, using %s", defaultVarFile)
 		}
 	}
 
 	mergedVariables := make(map[string]interface{})
-	for _, variableFile := range variableFiles {
+	for _, variableFile := range t.variableFiles {
 		// Process the variable file extension and log DEBUG so the template can be
 		// correctly rendered.
 		var ext string
@@ -117,18 +114,21 @@ func (t *tmpl) parseJSONVars(variableFile string) (variables map[string]interfac
 	return variables, nil
 }
 
-func (t *tmpl) parseTFVars(variableFile string) (variables map[string]interface{}, err error) {
+func (t *tmpl) parseTFVars(variableFile string) (map[string]interface{}, error) {
 
-	c := &config.Config{}
-	if c, err = config.LoadFile(variableFile); err != nil {
-		return
+	tfParser := configs.NewParser(nil)
+	loadedFile, loadDiags := tfParser.LoadConfigFile(variableFile)
+	if loadDiags != nil && loadDiags.HasErrors() {
+		return nil, loadDiags
+	}
+	if loadedFile == nil {
+		return nil, fmt.Errorf("hcl returned nil file")
 	}
 
-	variables = make(map[string]interface{})
-	for _, variable := range c.Variables {
-		variables[variable.Name] = variable.Default
+	variables := make(map[string]interface{})
+	for _, variable := range loadedFile.Variables {
+		variables[variable.Name] = hcl2shim.ConfigValueFromHCL2(variable.Default)
 	}
-
 	return variables, nil
 }
 
@@ -143,7 +143,6 @@ func (t *tmpl) parseYAMLVars(variableFile string) (variables map[string]interfac
 	if err = yaml.Unmarshal(yamlFile, &variables); err != nil {
 		return
 	}
-
 	return variables, nil
 }