+++ template/test-fixtures/test.tf
@@ -1,3 +1,9 @@
 variable "job_name" {
   default = "levantExample"
 }
+
+variable "task_resource_cpu" {
+  description = "the CPU for the task"
+  type        = number
+  default     = 1313
+}